Render call messages, break MessageMock up into multiple components
