+++ src/containers/message-mock/components/chained-date.tsx
@@ -1,0 +1,36 @@
+import { Box, Typography, useTheme } from "@mui/material";
+import { format, parseISO } from "date-fns";
+import Message from "../../../classes/message";
+
+type ChainedDateProps = { message: Message; longDateTime: string };
+
+const ChainedDate = ({ message, longDateTime }: ChainedDateProps) => {
+  const theme = useTheme();
+  const messageDate = parseISO(message.timestamp);
+  const shortTime = format(messageDate, "HH:mm:ss");
+
+  return (
+    <Box
+      title={longDateTime}
+      sx={{
+        width: "fit-content",
+        height: "20px",
+        display: "flex",
+        alignItems: "center",
+      }}
+    >
+      <Typography
+        id={`chained-message-${message.id}`}
+        sx={{
+          fontSize: "10px !important",
+          color: theme.palette.text.disabled,
+          opacity: 0,
+        }}
+        variant="caption"
+      >
+        {shortTime}
+      </Typography>
+    </Box>
+  );
+};
+export default ChainedDate;
